+++ src/common/js/configure.js
@@ -10,9 +10,6 @@ let _conf = {};
 EnvSettings._conf = new ReactiveVar( _conf );
 
 EnvSettings._defaults = {
-    reconfigurePackages: true,
-    sourcePath: 'environments',
-    targetPath: 'environment',
     verbosity: EnvSettings.C.Verbose.CONFIGURE
 };
 
@@ -28,9 +25,7 @@ EnvSettings.configure = function( o ){
         _conf = _.merge( EnvSettings._defaults, _conf, o );
         EnvSettings._conf.set( _conf );
         // verbosity management _after_ having set the conf
-        if( _conf.verbosity & EnvSettings.C.Verbose.CONFIGURE ){
-            console.log( 'pwix:env-settings configure() with', o );
-        }
+        EnvSettings.verbose( EnvSettings.C.Verbose.CONFIGURE, 'pwix:env-settings configure() with', o );
     }
     // also acts as a getter
     return EnvSettings._conf.get();