+++ src/common/js/constants.js
@@ -20,11 +20,6 @@ EnvSettings.C = {
         RECONFIGURE:  0x01 << 10
     },
 
-    // whether to wait for the Meteor startup() to load the configuration private assets
-    // this should be done by the application, but the package is initialized before the former
-    // starting with 2.0.0, we do not wait for startup, but is easy here to fallback to the historic behavior
-    WaitForStartup: false,
-
     // an internal publication to let the server propagates the 'ready' status to the clients
     readyPub: {
         publication: 'pwix:env-settings.ready.publication',