+++ src/common/js/index.js
@@ -7,6 +7,5 @@ import './constants.js';
 import './verbose.js';
 //
 import './configure.js';
-import './functions.js';
 import './ready.js';
 import './startup.js';